[CI skip] Added two more contributor names, some refactoring
