+++ src/main/java/me/mrCookieSlime/Slimefun/utils/Utilities.java
@@ -18,6 +18,8 @@ import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
 import org.bukkit.inventory.ItemStack;
 
+import io.github.thebusybiscuit.slimefun4.core.services.github.Contributor;
+import io.github.thebusybiscuit.slimefun4.core.services.github.GitHubConnector;
 import me.mrCookieSlime.CSCoreLibPlugin.Configuration.Config;
 import me.mrCookieSlime.Slimefun.GEO.OreGenResource;
 import me.mrCookieSlime.Slimefun.Lists.Categories;
@@ -39,8 +41,6 @@ import me.mrCookieSlime.Slimefun.api.item_transport.CargoTransportEvent;
 import me.mrCookieSlime.Slimefun.api.item_transport.ItemRequest;
 import me.mrCookieSlime.Slimefun.guides.ISlimefunGuide;
 import me.mrCookieSlime.Slimefun.guides.SlimefunGuideLayout;
-import me.mrCookieSlime.Slimefun.hooks.github.Contributor;
-import me.mrCookieSlime.Slimefun.hooks.github.GitHubConnector;
 
 /**
  * Really dirty way to store stuff, but you can dump
@@ -61,7 +61,7 @@ public final class Utilities {
 	public final List<MultiBlock> allMultiblocks = new LinkedList<>();
 	public final List<Research> allResearches = new LinkedList<>();
 	
-	public final Map<SlimefunGuideLayout, ISlimefunGuide> guideLayouts = new HashMap<>();
+	public final Map<SlimefunGuideLayout, ISlimefunGuide> guideLayouts = new EnumMap<>(SlimefunGuideLayout.class);
 	
 	/**
 	 * List of the registered Categories.
@@ -126,12 +126,12 @@ public final class Utilities {
 	public final Set<UUID> teleporterUsers = new HashSet<>();
 	
 	public final Map<String, OreGenResource> resources = new HashMap<>();
-	public final Map<String, Config> resource_configs = new HashMap<>();
+	public final Map<String, Config> resourceConfigs = new HashMap<>();
 	
 	public final Set<GitHubConnector> connectors = new HashSet<>();
 	public final ConcurrentMap<String, Contributor> contributors = new ConcurrentHashMap<>();
 	
-	public final List<CargoTransportEvent> cargoTransportEvents = new ArrayList<>();
+	public final Set<CargoTransportEvent> cargoTransportEvents = new HashSet<>();
 
 	public ItemStack[] oreWasherOutputs;
